version 0.0.6
in this version

Updated the documentation to make it easier for students to run them
removed migrations from the backend. better to create it fresh
there was just one connection string dealing with both auth db and app db. now there are two, one each for the db context
added a footer to the front end
added a abous us page to the front end
usual documentation updates