Update Program.cs
Added Code to check if Images folder does not exist. 

If images folder does not exist, it gets created. This is essential because I have gotten to the habit of not pushing the image folder (because sometimes, the images are personal and don't want it in a public repo), and simply deleting it after every development run. 

But, once the images folder is deleted, it was giving a run time error. Now that error has been removed.