Payment Flow Completed
Multiple improvements to the backend, and both angular and react front end to achieve a payment flow.

Users can now add a product to the shopping cart.
From the shopping cart, move to the payment page.
Make payment with a dummy payment page.
Once payment is done, confirm the order as well.

Still, plenty of admin views and user views related to orders and payments needs to be added, th basic flow is now in place.

Next, need to work on improvements but primary flow is looking good.