Added the User Info Table.
As described in the linked issue, I have added a user info table.
This is separate from the Microsoft Identity provided User Table.
This is like a custom table for extra information related to the user table.
When a new user registers, a new entry is created using the email address in the user table.
This makes sense because the idenity table also has email and that is linked to the email in the extra info table
However, I am keeping it a 'flat database' system where these two tables are linked but intentionally not doing any primary key and foreign key thing going on between them.
The User table can be edited from the admin panel.
Also, User can edit his own information in the profile page.